Add support for loading device_model and app_version from config file.

+++ scp/plugins/user/automations.py
@@ -18,6 +18,9 @@ __PLUGIN__ = 'automations'
     group=100,
 )
 async def auto_read_handler(_, message: Message):
+    if not user.scp_config.auto_read_mode:
+        return
+    
     if (not user.auto_read_enabled or message.id % 2 != 0 or
         message.chat.id == user.scp_config.avalon_pms):
         return
@@ -31,29 +34,26 @@ async def auto_read_handler(_, message: Message):
     user.owner & user.command('gautoread'),
 )
 async def gautoread_handler(_, message: Message):
+    if not user.scp_config.auto_read_mode:
+        return await message.reply_text('Auto read is disabled globally in config.')
+    
     if message.text.find('true') > 0:
         if user.auto_read_enabled:
-            await message.reply_text('Auto read is already enabled.')
-            return
-        user.auto_read_enabled = True
-        await message.reply_text('Auto read has been enabled.')
+            return await message.reply_text('Auto read is already enabled.')
         
-        return
+        user.auto_read_enabled = True
+        return await message.reply_text('Auto read has been enabled.')
 
     if message.text.find('false') > 0:
         if not user.auto_read_enabled:
-            await message.reply_text('Auto read is already disabled.')
-            return
+            return await message.reply_text('Auto read is already disabled.')
         user.auto_read_enabled = False
-        await message.reply_text('Auto read has been disabled.')
-        
-        return
+        return await message.reply_text('Auto read has been disabled.')
 
     if user.auto_read_enabled:
         await message.reply_text('Auto read is enabled.')
     else:
         await message.reply_text('Auto read is disabled.')
-    return
 
 
 @user.on_message(
@@ -63,12 +63,15 @@ async def gautoread_handler(_, message: Message):
     group=100,
 )
 async def auto_remove_bluetext_handler(_, message: Message):
-    if not user.auto_read_enabled or message.chat.title.lower().find('test') != -1:
+    if not user.scp_config.auto_read_mode:
+        return await message.reply_text('Auto read is disabled globally in config.')
+    
+    if not user.auto_read_enabled or \
+        message.chat.title.lower().find('test') != -1:
         return
     elif message.text.find('@') != -1: await asyncio.sleep(1.2)
+    
     try:
         await message.delete()
-    except Exception:
-        return
-    return
+    except Exception: return
